Set SameSite policy on the cookie
This seems to fix the issue that the gateway choice is "forgotten" from time to time, since the
cookie is rejected. Why this only occurs once in a while is not clear, but with the SameSite policy
I have not been able to trigger the issue locally anymor.

Also set some other cookie params while at it.